Improve syntax for artifact.sync config [1/3] (#1847)
* Improved pipeline config for artifact.sync

Currently, the pipeline config is a map of local glob pattern to destination directory.
This scheme has been extended with some magic sequences, making it difficult to understand.
This commit converts the sync map into a list of sync rules. All sync rules are consulted to determine the destination paths.

This prepares for further changes to the sync logic. Also see #1844
* Update sync spec config according to design proposal

- Wrap sync-rules under key 'sync.manual'
- Config changes:
- from -> src
- to -> dest
- flatten option removed and warn during schema upgrade if an incompatible pattern is migrated to the new schema version

* Add validation for sync rules
* Migrate sync config to new schema version
* Update filesync documentation

Extract example into sample snippet in order to have it tested.

* Run validation on doc examples
* Review comments: add test case and clarify example in docs
* Improve doc and error message wording

Signed-off-by: Cornelius Weig <22861411+corneliusweig@users.noreply.github.com>